+++ packages/amplify-ui-components/src/index.ts
@@ -1,5 +1,7 @@
-export { AuthState } from './common/types/auth-types';
+export * from './common/types/auth-types';
+export * from './common/types/ui-types';
 export { Translations } from './common/Translations';
+export { onAuthUIStateChange } from './common/helpers';
 export { AUTH_CHANNEL, UI_AUTH_CHANNEL, TOAST_AUTH_ERROR_EVENT, AUTH_STATE_CHANGE_EVENT } from './common/constants';
 export * from './components';
 export * from './components/amplify-auth-fields/amplify-auth-fields-interface';