+++ packages/aws-amplify-react-native/src/AmplifyUI.tsx
@@ -11,7 +11,7 @@
  * and limitations under the License.
  */
 
-import React, { Component } from 'react';
+import React, { Component, FC } from 'react';
 import {
 	Keyboard,
 	Picker,
@@ -23,9 +23,12 @@ import {
 	TouchableWithoutFeedback,
 	View,
 	SafeAreaView,
+	TextInputProperties,
+	TouchableOpacityProps,
 } from 'react-native';
 import { I18n } from 'aws-amplify';
 import AmplifyTheme, {
+	AmplifyThemeType,
 	linkUnderlayColor,
 	errorIconColor,
 	placeholderColor,
@@ -34,12 +37,22 @@ import { Icon } from 'react-native-elements';
 import countryDialCodes from './CountryDialCodes';
 import TEST_ID from './AmplifyTestIDs';
 
-export const Container = props => {
+interface IContainerProps {
+	theme?: AmplifyThemeType;
+}
+
+export const Container: FC<IContainerProps> = props => {
 	const theme = props.theme || AmplifyTheme;
 	return <SafeAreaView style={theme.container}>{props.children}</SafeAreaView>;
 };
 
-export const FormField = props => {
+interface IFormFieldProps extends TextInputProperties {
+	label: string;
+	required?: boolean;
+	theme?: AmplifyThemeType;
+}
+
+export const FormField: FC<IFormFieldProps> = props => {
 	const theme = props.theme || AmplifyTheme;
 	return (
 		<View style={theme.formField}>
@@ -57,8 +70,22 @@ export const FormField = props => {
 	);
 };
 
-export class PhoneField extends Component {
-	constructor(props) {
+interface IPhoneProps extends TextInputProperties {
+	defaultDialCode?: string;
+	label: string;
+	onChangeText: (phoneNumber: string) => void;
+	required?: boolean;
+	theme?: AmplifyThemeType;
+	value?: string;
+}
+
+interface IPhoneState {
+	dialCode: string;
+	phone: string;
+}
+
+export class PhoneField extends Component<IPhoneProps, IPhoneState> {
+	constructor(props: IPhoneProps) {
 		super(props);
 
 		this.state = {
@@ -120,29 +147,15 @@ export class PhoneField extends Component {
 	}
 }
 
-export const SectionFooter = props => {
-	const theme = props.theme || AmplifyTheme;
-	return (
-		<View style={theme.sectionFooter}>
-			<LinkCell
-				theme={theme}
-				onPress={() => onStateChange('confirmSignUp')}
-				testID={TEST_ID.AUTH.CONFIRM_A_CODE_BUTTON}
-			>
-				{I18n.get('Confirm a Code')}
-			</LinkCell>
-			<LinkCell
-				theme={theme}
-				onPress={() => onStateChange('signIn')}
-				testID={TEST_ID.AUTH.SIGN_IN_BUTTON}
-			>
-				{I18n.get('Sign In')}
-			</LinkCell>
-		</View>
-	);
-};
+interface ILinkCellProps {
+	disabled?: boolean;
+	onPress: () => void;
+	testID?: string;
+	theme?: AmplifyThemeType;
+}
 
-export const LinkCell = props => {
+export const LinkCell: FC<ILinkCellProps> = props => {
+	const { disabled } = props;
 	const theme = props.theme || AmplifyTheme;
 	return (
 		<View style={theme.cell}>
@@ -150,14 +163,26 @@ export const LinkCell = props => {
 				onPress={props.onPress}
 				underlayColor={linkUnderlayColor}
 				testID={props.testID}
+				disabled={disabled}
 			>
-				<Text style={theme.sectionFooterLink}>{props.children}</Text>
+				<Text
+					style={
+						disabled ? theme.sectionFooterLinkDisabled : theme.sectionFooterLink
+					}
+				>
+					{props.children}
+				</Text>
 			</TouchableHighlight>
 		</View>
 	);
 };
 
-export const Header = props => {
+interface IHeaderProps {
+	testID?: string;
+	theme?: AmplifyThemeType;
+}
+
+export const Header: FC<IHeaderProps> = props => {
 	const theme = props.theme || AmplifyTheme;
 	return (
 		<View style={theme.sectionHeader}>
@@ -168,7 +193,11 @@ export const Header = props => {
 	);
 };
 
-export const ErrorRow = props => {
+interface IErrorRowProps {
+	theme?: AmplifyThemeType;
+}
+
+export const ErrorRow: FC<IErrorRowProps> = props => {
 	const theme = props.theme || AmplifyTheme;
 	if (!props.children) return null;
 	return (
@@ -181,7 +210,14 @@ export const ErrorRow = props => {
 	);
 };
 
-export const AmplifyButton = props => {
+interface IAmplifyButtonProps extends TouchableOpacityProps {
+	disabled?: boolean;
+	style?: object;
+	text: string;
+	theme?: AmplifyThemeType;
+}
+
+export const AmplifyButton: FC<IAmplifyButtonProps> = props => {
 	const theme = props.theme || AmplifyTheme;
 	let style = theme.button;
 	if (props.disabled) {
@@ -199,11 +235,19 @@ export const AmplifyButton = props => {
 	);
 };
 
-export const Wrapper = props => {
-	const isWeb = Platform.OS === 'web';
-	const WrapperComponent = isWeb ? View : TouchableWithoutFeedback;
+interface IWrapperProps {
+	style?: AmplifyThemeType;
+	accessible?: boolean;
+	onPress?: Function;
+}
 
-	const wrapperProps = {
+export const Wrapper: FC<IWrapperProps> = props => {
+	const isWeb = Platform.OS === 'web';
+	const WrapperComponent: React.ElementType = isWeb
+		? View
+		: TouchableWithoutFeedback;
+
+	const wrapperProps: IWrapperProps = {
 		style: AmplifyTheme.section,
 		accessible: false,
 	};
@@ -214,5 +258,19 @@ export const Wrapper = props => {
 
 	return (
 		<WrapperComponent {...wrapperProps}>{props.children}</WrapperComponent>
+	);
+};
+
+export const SignedOutMessage = props => {
+	const theme = props.theme || AmplifyTheme;
+	const message =
+		props.signedOutMessage || I18n.get('Please Sign In / Sign Up');
+	return (
+		<Text
+			style={theme.signedOutMessage}
+			testID={TEST_ID.AUTH.GREETING_SIGNED_OUT_TEXT}
+		>
+			{message}
+		</Text>
 	);
 };